refactor: extract config file search to shared utility
Move duplicated config file search logic from linter.ts and fixer.ts
to a shared resolveStandard() function in linter-utils.ts.

- Add CONFIG_FILE_NAMES constant for config file precedence
- Add resolveStandard() utility function
- Update both linter.ts and fixer.ts to use the shared utility
- Remove unused extfs import from fixer.ts
